category/requirements.txt
Flask==3.1.0 
requests==2.28.0
Flask-MySQLdb==2.0.0
pandas==2.2.3
transformers
TensorFlow==2.19.0
torch
tf_keras
transformers
sentence-transformers
pyyaml